changed route for create and moved validate function to utils
