+++ api/index.js
@@ -3,7 +3,7 @@ const { conn } = require("./src/db.js");
 const port = process.env.PORT || 3001;
 require("./src/services/PaymentService");
 
-conn.sync({ alter: true }).then(() => {
+conn.sync({ force: true }).then(() => {
   server.listen(3001, () => {
     console.log(`%s listening at ${port}`); // eslint-disable-line no-console
   });
